// src/utils/supabase/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)";
  };
  public: {
    Tables: {
      Box: {
        Row: {
          created_at: string;
          id: string;
          name: string;
          password_hash: string | null;
          password_protected: boolean;
        };
        Insert: {
          created_at?: string;
          id?: string;
          name: string;
          password_hash?: string | null;
          password_protected?: boolean;
        };
        Update: {
          created_at?: string;
          id?: string;
          name?: string;
          password_hash?: string | null;
          password_protected?: boolean;
        };
        Relationships: [];
      };
      FileContent: {
        Row: {
          box: string;
          content: string;
          created_at: string;
          hide_content: boolean;
          id: string;
        };
        Insert: {
          box?: string;
          content?: string;
          created_at?: string;
          hide_content?: boolean;
          id?: string;
        };
        Update: {
          box?: string;
          content?: string;
          created_at?: string;
          hide_content?: boolean;
          id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "FileContent_box_fkey";
            columns: ["box"];
            isOneToOne: false;
            referencedRelation: "Box";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "FileContent_box_fkey";
            columns: ["box"];
            isOneToOne: false;
            referencedRelation: "PublicBox";
            referencedColumns: ["id"];
          },
        ];
      };
      ImageContent: {
        Row: {
          box: string;
          content: string;
          created_at: string;
          hide_content: boolean;
          id: string;
        };
        Insert: {
          box?: string;
          content?: string;
          created_at?: string;
          hide_content?: boolean;
          id?: string;
        };
        Update: {
          box?: string;
          content?: string;
          created_at?: string;
          hide_content?: boolean;
          id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "ImageContent_box_fkey";
            columns: ["box"];
            isOneToOne: false;
            referencedRelation: "Box";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "ImageContent_box_fkey";
            columns: ["box"];
            isOneToOne: false;
            referencedRelation: "PublicBox";
            referencedColumns: ["id"];
          },
        ];
      };
      TextContent: {
        Row: {
          box: string;
          content: string;
          created_at: string;
          hide_content: boolean;
          id: string;
        };
        Insert: {
          box?: string;
          content?: string;
          created_at?: string;
          hide_content?: boolean;
          id?: string;
        };
        Update: {
          box?: string;
          content?: string;
          created_at?: string;
          hide_content?: boolean;
          id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "TextContent_box_fkey";
            columns: ["box"];
            isOneToOne: false;
            referencedRelation: "Box";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "TextContent_box_fkey";
            columns: ["box"];
            isOneToOne: false;
            referencedRelation: "PublicBox";
            referencedColumns: ["id"];
          },
        ];
      };
    };
    Views: {
      PublicBox: {
        Row: {
          created_at: string | null;
          id: string | null;
          name: string | null;
          password_protected: boolean | null;
        };
        Insert: {
          created_at?: string | null;
          id?: string | null;
          name?: string | null;
          password_protected?: boolean | null;
        };
        Update: {
          created_at?: string | null;
          id?: string | null;
          name?: string | null;
          password_protected?: boolean | null;
        };
        Relationships: [];
      };
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  "public"
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
  ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
  : never;

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
  ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
  : never;